+++ R/fit_SE.combinations.R
@@ -112,7 +112,8 @@ fit_SE.combinations <- function(se,
       "rconcs"
     )
     
-    mean_avg_combo <-  avg_combo[, lapply(.SD, mean), by = c(id, id2, "normalization_type"), .SDcols = "x"]
+    mean_avg_combo <-  avg_combo[, lapply(.SD, mean), by = c(id, id2, "normalization_type"), .SDcols = c("x", "x_std")]
+
     # deal with cases of multiple concentrations mapped to the same value 
     # when rounded create a complete matrix with the most frequence combo 
     # concentrations
@@ -140,7 +141,9 @@ fit_SE.combinations <- function(se,
   
     for (norm_type in normalization_types) {
 
-      avg_combo <- data.table::as.data.table(avg_combo)
+      # use data.table with averaged concentration values 
+      # (if some duplicates appeared after mapping to standardized concentrations)
+      avg_combo <- data.table::as.data.table(mean_avg_combo)
       avg_subset <- avg_combo[normalization_type == norm_type]
       complete_subset <- complete[normalization_type == norm_type | is.na(normalization_type)]
       complete_subset[is.na(normalization_type), normalization_type := norm_type]